data_scientist_part1/requirements.txt
pandas==2.2.3
psycopg2-binary==2.9.10
numpy==2.2.5
matplotlib==3.10.1
scipy==1.15.2